+++ app/src/components/Announcement/index.tsx
@@ -1,6 +1,5 @@
 import mountains from '@/assets/mountains.png';
 import wordLogo from '@/assets/word-logo.svg';
-import supabase from '@/lib/supabase';
 import { DialogTitle } from '@radix-ui/react-dialog';
 import {
     BoxIcon,
@@ -18,6 +17,7 @@ import { toast } from '../ui/use-toast';
 import { AuthButton } from './auth';
 import { Links, MainChannels } from '/common/constants';
 import { UserSettings } from '/common/models/settings';
+import supabase from '/common/supabase';
 
 function Announcement() {
     const [checked, setChecked] = useState(true);
