--- app/src/lib/supabase/index.ts
@@ -1,14 +1,0 @@
-import { createClient } from '@supabase/supabase-js';
-
-let supabase: ReturnType<typeof createClient> | undefined;
-
-try {
-    supabase = createClient(
-        import.meta.env.VITE_SUPABASE_API_URL || '',
-        import.meta.env.VITE_SUPABASE_ANON_KEY || '',
-    );
-} catch (error) {
-    console.error('Error initializing Supabase:', error);
-}
-
-export default supabase;
